@@ -10,21 +10,22 @@
   "author": "Vitaliy Otroshenko",
   "license": "ISC",
   "dependencies": {
-    "@types/node": "^16.9.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "styled-components": "^5.3.1",
-    "typescript": "^4.4.3"
+    "@types/node": "^18.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "styled-components": "^5.3.5"
   },
   "devDependencies": {
-    "@parcel/transformer-image": "2.0.0-rc.0",
-    "@parcel/transformer-sass": "2.0.0-rc.0",
-    "@types/react": "^17.0.21",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.1.9",
-    "@types/styled-components": "^5.1.14",
-    "parcel": "^2.0.0-rc.0",
-    "prettier": "^2.4.1"
+    "@parcel/transformer-image": "^2.6.2",
+    "@parcel/transformer-sass": "^2.6.2",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/styled-components": "^5.1.25",
+    "parcel": "^2.6.2",
+    "prettier": "^2.7.1",
+    "process": "^0.11.10",
+    "typescript": "^4.7.4"
   }
 }
