@@ -3,7 +3,7 @@ import React, {useState} from "react";
 import {LoginPageWrapper, LoginFormWrapper, LoginUserAvatar, LoginError, Avatar} from "./LoginPage.styled";
 import {Input} from "../../components/inputs/Input";
 import {LoginButton} from "../../components/button/LoginButton";
-import {useHistory} from "react-router-dom";
+import {useNavigate} from "react-router-dom";
 import userAvatar from "./assets/userAvatar.png";
 
 export function LoginPage() {
@@ -11,7 +11,7 @@ export function LoginPage() {
   const [password, setPassword] = useState<string>("");
   const [error, setError] = useState<string | null>(null);
 
-  const history = useHistory();
+  const history = useNavigate();
 
   const logIn = () => {
     if (!userName || !password) {
@@ -19,9 +19,8 @@ export function LoginPage() {
       return;
     }
 
-    localStorage.setItem("auth", "success");
     setError(null);
-    history.push("/layout");
+    history("/layout");
   };
 
   return (
@@ -36,7 +35,9 @@ export function LoginPage() {
         <Input value={userName} onChange={setUserName} label="User Name" />
         <Input type="password" value={password} onChange={setPassword} label="Password" />
 
-        <LoginButton onClick={logIn}>Log in</LoginButton>
+        <LoginButton type="button" onClick={logIn}>
+          Log in
+        </LoginButton>
       </LoginFormWrapper>
     </LoginPageWrapper>
   );
